fix : 다른 month로 이동해도 date가 유지되는 버그 수정
